requirements.txt
Flask==3.0.3
Flask_RESTful==0.3.10
mysql_connector_repackaged==0.3.1
pydantic==2.7.1
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
Requests==2.31.0
utils==1.0.2
waitress==3.0.0
